CLOSED - task 108: Add tag cloud http://github.com/mPowering/django-mpowering/issues/issue/108

+++ mpowering/templatetags/display_functions.py
@@ -4,4 +4,4 @@ register = template.Library()
 
 @register.filter(name='cloud_text_size')
 def cloud_text_size(count, diff):
-    return 10 + int((40/(diff))*count)
+    return 10 + ((40/(diff))*count)
